+++ src/Pages/LandingPage/BannerRegister/styles.scss
@@ -1,6 +1,5 @@
 @import '../../../assets/scss/variable';
 @import '../../../assets/scss/breakpoints';
-
 .sponser {
   position: relative;
   padding-top: 70px;
