--- src/Pages/LandingPage/LandingPage.js
@@ -1,13 +1,0 @@
-import React, { Component, Fragment } from 'react';
-import Header from './../../components/Header/Header'
-class LandingPage extends Component {
-    render() {
-        return (
-            <Fragment>
-                <Header/>
-            </Fragment>
-        );
-    }
-}
-
-export default LandingPage;
