+++ src/Pages/LandingPage/Participation/styles.scss
@@ -1,5 +1,8 @@
 @import '../../../assets/scss/variable';
 
+.bg-w {
+  background: #fff !important;
+}
 .participation {
   position: relative;
   background-color: #fff;
